cleaned newton folder and renamed spec
