+++ Vol1B/NewtonsMethod/oldlab/newton.py
