+++ Vol1B/NewtonsMethod/oldlab/plots2.py
