+++ Vol1B/NewtonsMethod/oldlab/solutionsFull.py
