+++ Vol1B/NewtonsMethod/oldlab/spec.py
