+++ Vol1B/NewtonsMethod/oldlab/zoom_j.py
